Support external sources - can add, remove, edit existing source as well as provide new / delete
Added middlewareCollection.checkValidityOfPenId to prevent pen page iFrame from calling invalid external source urls when rendering code from editors/external sources
Changed externalString (used by render) generation to instead use an externalsDictionary rather than from the server provided object
Added deletion capability for any external source - an external source marked with 'delete' will be ignored for rendering and will be removed on a 'save' operation
Update returnRenderContentForiFrame to include external sources
Update post/put operations to server to include penExternals
Do an initial render on page load by adding refreshRenderContent to the monaco_initializeEditors function which loads in monaco dependencies - calling this earlier will cause issues as monaco is loaded in via require, doing this on $.ready won't work
Added additional variables to enable modal functionality for dynamically creating new list items to display existing css/js sources
Support adding temporary IDs to new css/js sources; on a successful PUT, the penExternals object will contain the IDs generated by the server/DB
Add basic validation to the css/js source to prevent a save if the provided input is not a url
Update the penExternals object (rewrite) with the contents maintained in the externalsDictionary - the dictionary is created on page load and is updated on any change to any specific externals line item
Add a function for updating externalsDictionary content when the 'save/exit' selection is made in the modal; trigger re-render as well
Update modal links on the html/css/js UI to open the respective tabs in the modal
Add a deletion function on the modal externals list to perform a 'soft' delete on the client side; hard delete is done when a save is triggered, causing both externalsDictionary and penExternals to refresh contents
Update server pen POST/PUT to support penExternals. For PUT, 3 cases are handled - if the source exists, it is updated. if the source is marked for deletion, it is deleted, if it is a new source, a new row is inserted.